F14: Store the humidity of a given tile as a float

+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a2/island/humidity/IHumidity.java
@@ -5,13 +5,13 @@ public interface IHumidity {
      *
      * @return The humidity belonging to this element
      */
-    double getHumidity();
+    float getHumidity();
 
     /**
      *
      * @param humidity The humidity to set this element to
      */
-    void setHumidity(double humidity);
+    void setHumidity(float humidity);
 
     /**
      *
