+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a2/island/humidity/handlers/transmitter/HumidityTransmitter.java
@@ -10,7 +10,7 @@ public class HumidityTransmitter implements IHumidityTransmitter {
     }
 
     @Override
-    public double giveHumidity(IHumidity humidity) {
+    public float giveHumidity(IHumidity humidity) {
         return this.ratio * humidity.getHumidity();
     }
 }
