+++ island/src/main/java/ca/mcmaster/cas/se2aa4/a2/island/humidity/handlers/transmitter/IHumidityTransmitter.java
@@ -8,5 +8,5 @@ public interface IHumidityTransmitter {
      * @param humidity The humidity element that is giving off humidity
      * @return The humidity given off by the element
      */
-    double giveHumidity(IHumidity humidity);
+    float giveHumidity(IHumidity humidity);
 }
